fix hw2
