+++ src/main/java/ru/vk/itmo/kovalevigor/MemoryEntryWriter.java
@@ -33,12 +33,4 @@ public class MemoryEntryWriter extends MemoryEntryFileWorker {
         putMemorySegment(entry.getValue().value());
     }
 
-    public static long getTotalMapSize(final Map<MemorySegment, Entry<MemorySegment>> map) {
-        long totalSize = ENTRY_META_SIZE * map.size();
-        for (Map.Entry<MemorySegment, Entry<MemorySegment>> entry : map.entrySet()) {
-            totalSize += entry.getKey().byteSize() + entry.getValue().value().byteSize();
-        }
-        return totalSize;
-    }
-
 }
